+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## [0.25.1] - 2024-03-21
+
+- Prevent SingleDB mode from asking for `--destroy-without-prefix` CLI option which is not possible to set on "destroy" action.
+- Make sure schemas are correctly "destroyed" even when `DatabaseResolver` is not present in resolver sequence. Most schema objects are still being ignored.
+
 ## [0.25.0] - 2024-03-20
 
 - Added browser-based SSO authentication (thanks to Joseph Niblo).